app/requirement.txt
streamlit
numpy
pillow
scikit-learn
streamlit-drawable-canvas